physics now doubles ... update and draw loops missing atm

+++ 3dplayground/Physics/Interfaces/IFieldPhysics.cs
@@ -12,7 +12,7 @@ namespace _3dplayground.Physics
     /// </summary>
     interface IFieldPhysics
     {
-        New_pos_and_vel dothe_phys(float step, IGetEffectedByGravity i);
+        New_pos_and_vel dothe_phys(double step, IGetEffectedByGravity i);
         //ToDo: Rick: i have no idea what imput parameters you need or the return type. 
         //If you need to return an args object so be it - although output params maybe ...
         // anyway ... returning a some kind of a vector would be nice.
